Fix issue with XML output tests running in parallel. Add JackCompiler section to README.md.
